reflection: widen large-uint pad arithmetic to int64 (CodeQL)
The spec-derived uint32 size cannot be bounded by the compiler, so use int64
for the pad computation to avoid a truncating uint32->int conversion on 32-bit
platforms.

+++ reflection/treeroot.go
@@ -365,7 +365,10 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 	} else {
 		b := sourceValue.Bytes()
 		hh.Append(b)
-		if pad := int(sourceType.Size) - len(b); pad > 0 {
+		// int64 keeps the widening conversion from the uint32 size safe on every
+		// platform (a large-uint width is only ever 16 or 32, but the size is a
+		// spec-derived uint32 the compiler cannot bound).
+		if pad := int64(sourceType.Size) - int64(len(b)); pad > 0 {
 			hh.Append(sszutils.ZeroBytes()[:pad])
 		}
 	}
